removed warning.
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@1563 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/function/adaptivefunction/adaptivefunction.hh
@@ -623,7 +623,7 @@ namespace Dune {
         for (SizeType j = 0; j < N; ++j )
         {
           // only works for point base mappings 
-          assert( (idx+j) == spc.mapToGlobal(en, i*N+j) );
+          assert( (idx+ (int) j) == spc.mapToGlobal(en, i*N+j) );
           values[i][j] = &(dofVec[ idx+j ]);
         } // end for j
       }
